requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.0
idna==3.7
inflection==0.5.1
mysqlclient==2.2.4
numpy==1.26.4
pygini==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
